requirements.txt
# Django and Django REST Framework
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-yasg==1.21.7

# Environment Variables
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# Image Processing
Pillow==10.1.0

# Production Server
gunicorn==21.2.0

# Database Migrations
django-extensions==3.2.3

# Filtering
django-filter==23.5

# Full-text Search (PostgreSQL)
django-postgres-extra==2.0.0

# Utilities
python-dateutil==2.8.2

# Development Tools (Optional)
ipython==8.18.1
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Logging & Monitoring
sentry-sdk==1.38.0
python-json-logger==2.0.7
psutil==5.9.6

# Caching
redis==5.0.1
django-redis==5.4.0

# Rate Limiting
django-ratelimit==4.1.0

# File Storage & Processing
storages==1.14.2
boto3==1.34.0
supabase==2.0.3
Pillow==10.1.0
pdf2image==1.16.3
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
ClamAV==1.0.1
django-cleanup==8.0.0

# Search Enhancements
django-extensions==3.2.3

# Background Tasks
celery==5.3.4
celery-beat==2.5.0
redis==5.0.1

# Push Notifications (Optional)
# pyfcm==1.5.5  # Firebase Cloud Messaging
# onesignal-sdk==1.2.0  # OneSignal